added time in same line
