+++ cmd/withdraw_test.go
@@ -36,6 +36,7 @@ func TestWithdrawFunds(t *testing.T) {
 		txnOpts                  *bind.TransactOpts
 		epoch                    uint32
 		epochErr                 error
+		time                     string
 		withdrawHash             common.Hash
 		withdrawErr              error
 	}
@@ -160,6 +161,7 @@ func TestWithdrawFunds(t *testing.T) {
 				txnOpts:                  txnOpts,
 				epoch:                    3,
 				epochErr:                 nil,
+				time:                     "10 minutes 0 seconds ",
 				withdrawHash:             common.BigToHash(big.NewInt(1)),
 				withdrawErr:              nil,
 			},
@@ -201,6 +203,7 @@ func TestWithdrawFunds(t *testing.T) {
 			utilsMock.On("GetTxnOpts", mock.AnythingOfType("types.TransactionOptions")).Return(txnOpts)
 			utilsMock.On("GetEpoch", mock.AnythingOfType("*ethclient.Client")).Return(tt.args.epoch, tt.args.epochErr)
 			utilsMock.On("CalculateBlockTime", mock.AnythingOfType("*ethclient.Client")).Return(blockTime)
+			utilsMock.On("SecondsToHuman", mock.AnythingOfType("int")).Return(tt.args.time)
 			cmdUtilsMock.On("Withdraw", mock.Anything, mock.Anything, mock.Anything).Return(tt.args.withdrawHash, tt.args.withdrawErr)
 
 			utils := &UtilsStruct{}
@@ -406,67 +409,6 @@ func TestExecuteWithdraw(t *testing.T) {
 				t.Error("The ExecuteWithdraw function didn't execute as expected")
 			}
 
-		})
-	}
-}
-
-func TestSecondsToHuman(t *testing.T) {
-	type args struct {
-		input int
-	}
-	tests := []struct {
-		name       string
-		args       args
-		wantResult string
-	}{
-		{
-			name: "Test 1",
-			args: args{
-				input: 100,
-			},
-			wantResult: "1 minute 40 seconds ",
-		},
-		{
-			name: "Test 2",
-			args: args{
-				input: 10000000000,
-			},
-			wantResult: "45 years 11 months 4 weeks 2 days 17 hours 46 minutes 40 seconds ",
-		},
-		{
-			name: "Test 3",
-			args: args{
-				input: 100000000,
-			},
-			wantResult: "5 months 15 weeks 2 days 9 hours 46 minutes 40 seconds ",
-		},
-		{
-			name: "Test 4",
-			args: args{
-				input: 1000000,
-			},
-			wantResult: "1 week 4 days 13 hours 46 minutes 40 seconds ",
-		},
-		{
-			name: "Test 5",
-			args: args{
-				input: 100000,
-			},
-			wantResult: "1 day 3 hours 46 minutes 40 seconds ",
-		},
-		{
-			name: "Test 6",
-			args: args{
-				input: 10000,
-			},
-			wantResult: "2 hours 46 minutes 40 seconds ",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if gotResult := secondsToHuman(tt.args.input); gotResult != tt.wantResult {
-				t.Errorf("secondsToHuman() = %v, want %v", gotResult, tt.wantResult)
-			}
 		})
 	}
 }
