+++ utils/mocks/utils.go
@@ -1629,6 +1629,20 @@ func (_m *Utils) SaveDataToFile(_a0 string, _a1 uint32, _a2 []*big.Int) error {
 	return r0
 }
 
+// SecondsToHuman provides a mock function with given fields: _a0
+func (_m *Utils) SecondsToHuman(_a0 int) string {
+	ret := _m.Called(_a0)
+
+	var r0 string
+	if rf, ok := ret.Get(0).(func(int) string); ok {
+		r0 = rf(_a0)
+	} else {
+		r0 = ret.Get(0).(string)
+	}
+
+	return r0
+}
+
 // SuggestGasPriceWithRetry provides a mock function with given fields: _a0
 func (_m *Utils) SuggestGasPriceWithRetry(_a0 *ethclient.Client) (*big.Int, error) {
 	ret := _m.Called(_a0)
